Extract cross-cutting models from swatch-core into a smaller library
Carving out some of the models that are duplicated in swatch-contracts from swatch-core that have cross cutting concerns across multiple swatch services.  Starting with HardwareMeasurementType, ServiceLevel, and Usage - because we are going to need them in swatch-metrics-hbi.

This is just meant as a starting point and to unblock the swatch-metrics-hbi work.  In the future, we should try to make these schema-driven.  We should also have a conversation about if we want to split each model out into its own gradle module or not - in case a swatch services only share a subset of models.  They also need to be yanked from swatch-core as well but that got very sticky very fast.
