+++ settings.gradle
@@ -59,3 +59,4 @@ include ':swatch-common-export'
 include ':swatch-metrics'
 include ':swatch-common-trace-response'
 include ':swatch-metrics-hbi'
+include ':swatch-common-models'
