+++ swatch-common-models/src/main/java/com/redhat/swatch/common/model/HardwareMeasurementType.java
@@ -18,7 +18,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.redhat.swatch.contract.repository;
+package com.redhat.swatch.common.model;
 
 import java.util.Arrays;
 import java.util.List;
