+++ swatch-contracts/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation project(':swatch-common-smallrye-fault-tolerance')
     implementation project(':swatch-common-splunk')
     implementation project(':swatch-common-export')
+    implementation project(':swatch-common-models')
     implementation project(':clients:quarkus:rbac-client')
     implementation project(':clients:quarkus:product-client')
     implementation project(':clients:quarkus:export-client')
