+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/product/UpstreamProductData.java
@@ -30,8 +30,8 @@ import com.redhat.swatch.contract.exception.ErrorCode;
 import com.redhat.swatch.contract.exception.ServiceException;
 import com.redhat.swatch.contract.product.umb.UmbOperationalProduct;
 import com.redhat.swatch.contract.repository.OfferingEntity;
-import com.redhat.swatch.contract.repository.ServiceLevel;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.ServiceLevel;
+import com.redhat.swatch.common.model.Usage;
 import jakarta.ws.rs.core.Response;
 import java.util.ArrayList;
 import java.util.Arrays;
