+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/repository/Granularity.java
@@ -20,6 +20,7 @@
  */
 package com.redhat.swatch.contract.repository;
 
+import com.redhat.swatch.common.model.StringValueEnum;
 import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
