+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/repository/SubscriptionCapacityView.java
@@ -20,6 +20,8 @@
  */
 package com.redhat.swatch.contract.repository;
 
+import com.redhat.swatch.common.model.ServiceLevel;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.model.SubscriptionCapacityViewMetricConverter;
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
