+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/repository/SubscriptionCapacityViewRepository.java
@@ -26,6 +26,8 @@ import static com.redhat.swatch.contract.resource.ResourceUtils.sanitizeBillingP
 import static com.redhat.swatch.contract.resource.ResourceUtils.sanitizeServiceLevel;
 import static com.redhat.swatch.contract.resource.ResourceUtils.sanitizeUsage;
 
+import com.redhat.swatch.common.model.ServiceLevel;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.configuration.registry.MetricId;
 import com.redhat.swatch.configuration.registry.ProductId;
 import com.redhat.swatch.panache.PanacheSpecificationSupport;
