+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/resource/api/v1/CapacityResourceV1.java
@@ -35,12 +35,12 @@ import com.redhat.swatch.contract.openapi.model.UsageType;
 import com.redhat.swatch.contract.openapi.resource.CapacityApi;
 import com.redhat.swatch.contract.repository.DbReportCriteria;
 import com.redhat.swatch.contract.repository.Granularity;
-import com.redhat.swatch.contract.repository.HardwareMeasurementType;
+import com.redhat.swatch.common.model.HardwareMeasurementType;
 import com.redhat.swatch.contract.repository.HypervisorReportCategory;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionEntity;
 import com.redhat.swatch.contract.repository.SubscriptionRepository;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.resource.InMemoryPager;
 import com.redhat.swatch.contract.resource.ResourceUtils;
 import com.redhat.swatch.contract.resteasy.Page;
