+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/resource/api/v1/SubscriptionTableControllerV1.java
@@ -40,13 +40,13 @@ import com.redhat.swatch.contract.openapi.model.SubscriptionType;
 import com.redhat.swatch.contract.openapi.model.UsageType;
 import com.redhat.swatch.contract.repository.BillingProvider;
 import com.redhat.swatch.contract.repository.DbReportCriteria;
-import com.redhat.swatch.contract.repository.HardwareMeasurementType;
+import com.redhat.swatch.common.model.HardwareMeasurementType;
 import com.redhat.swatch.contract.repository.HypervisorReportCategory;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionEntity;
 import com.redhat.swatch.contract.repository.SubscriptionMeasurementKey;
 import com.redhat.swatch.contract.repository.SubscriptionRepository;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.resource.InMemoryPager;
 import io.quarkus.panache.common.Sort;
 import jakarta.enterprise.context.ApplicationScoped;
