+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/resource/api/v2/SubscriptionTableControllerV2.java
@@ -36,12 +36,12 @@ import com.redhat.swatch.contract.openapi.model.SubscriptionEventType;
 import com.redhat.swatch.contract.openapi.model.SubscriptionType;
 import com.redhat.swatch.contract.openapi.model.UsageType;
 import com.redhat.swatch.contract.repository.BillingProvider;
-import com.redhat.swatch.contract.repository.HardwareMeasurementType;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.HardwareMeasurementType;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionCapacityView;
 import com.redhat.swatch.contract.repository.SubscriptionCapacityViewMetric;
 import com.redhat.swatch.contract.repository.SubscriptionCapacityViewRepository;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.resource.InMemoryPager;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
