+++ swatch-contracts/src/main/java/com/redhat/swatch/contract/service/export/SubscriptionJsonDataMapperService.java
@@ -23,9 +23,9 @@ package com.redhat.swatch.contract.service.export;
 import static com.redhat.swatch.contract.service.export.SubscriptionDataExporterService.groupMetrics;
 
 import com.redhat.swatch.contract.model.SubscriptionsExportJsonItem;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionCapacityView;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.resource.api.v1.ApiModelMapperV1;
 import com.redhat.swatch.export.DataMapperService;
 import com.redhat.swatch.export.ExportServiceRequest;
