+++ swatch-contracts/src/test/java/com/redhat/swatch/contract/product/UpstreamProductDataTest.java
@@ -35,8 +35,8 @@ import com.redhat.swatch.contract.product.umb.ProductAttribute;
 import com.redhat.swatch.contract.product.umb.ProductRelationship;
 import com.redhat.swatch.contract.product.umb.UmbOperationalProduct;
 import com.redhat.swatch.contract.repository.OfferingEntity;
-import com.redhat.swatch.contract.repository.ServiceLevel;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.ServiceLevel;
+import com.redhat.swatch.common.model.Usage;
 import java.util.Collections;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
