+++ swatch-contracts/src/test/java/com/redhat/swatch/contract/resource/api/v1/CapacityResourceV1Test.java
@@ -40,11 +40,11 @@ import com.redhat.swatch.contract.repository.DbReportCriteria;
 import com.redhat.swatch.contract.repository.Granularity;
 import com.redhat.swatch.contract.repository.HypervisorReportCategory;
 import com.redhat.swatch.contract.repository.OfferingEntity;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionEntity;
 import com.redhat.swatch.contract.repository.SubscriptionMeasurementKey;
 import com.redhat.swatch.contract.repository.SubscriptionRepository;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import io.quarkus.panache.common.Sort;
 import io.quarkus.security.ForbiddenException;
 import io.quarkus.test.InjectMock;
