+++ swatch-contracts/src/test/java/com/redhat/swatch/contract/resource/api/v1/SubscriptionTableControllerV1Test.java
@@ -50,11 +50,11 @@ import com.redhat.swatch.contract.repository.BillingProvider;
 import com.redhat.swatch.contract.repository.DbReportCriteria;
 import com.redhat.swatch.contract.repository.HypervisorReportCategory;
 import com.redhat.swatch.contract.repository.OfferingEntity;
-import com.redhat.swatch.contract.repository.ServiceLevel;
+import com.redhat.swatch.common.model.ServiceLevel;
 import com.redhat.swatch.contract.repository.SubscriptionEntity;
 import com.redhat.swatch.contract.repository.SubscriptionMeasurementKey;
 import com.redhat.swatch.contract.repository.SubscriptionRepository;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.Usage;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
