+++ swatch-contracts/src/test/java/com/redhat/swatch/contract/service/OfferingSyncServiceTest.java
@@ -46,8 +46,8 @@ import com.redhat.swatch.contract.model.OfferingSyncTask;
 import com.redhat.swatch.contract.model.SyncResult;
 import com.redhat.swatch.contract.repository.OfferingEntity;
 import com.redhat.swatch.contract.repository.OfferingRepository;
-import com.redhat.swatch.contract.repository.ServiceLevel;
-import com.redhat.swatch.contract.repository.Usage;
+import com.redhat.swatch.common.model.ServiceLevel;
+import com.redhat.swatch.common.model.Usage;
 import com.redhat.swatch.contract.test.resources.ProductUseStubService;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
